+++ R/utils.R
@@ -137,15 +137,14 @@ get_time_resolution <- function(chrom_list, index=1){
 #' @noRd
 check_for_pkg <- function(pkg, return_boolean = FALSE){
   pkg_exists <- requireNamespace(pkg, quietly = TRUE)
-  if (!pkg_exists) {
+  if (return_boolean){
+    return(pkg_exists)
+  } else if (!pkg_exists) {
     stop(paste(
       "Package", sQuote(pkg), "must be installed to perform this action:
           try", paste0("`install.packages('", pkg, "')`.")),
       call. = FALSE
     )
-  }
-  if (return_boolean){
-    pkg_exists
   }
 }